+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-07-14
+**BREAKING RELEASE - needs a database migration:**
+* Before upgrade 
+```bash
+the-way export > snippets.json
+the-way clear
+```
+* After upgrade
+```bash
+the-way import snippets.json
+```
+### Added
+**Sync to Gist functionality!** [Issue #60](https://github.com/out-of-cheese-error/the-way/issues/60)
+### Fixed
+* `export filename` and `config default filename` work without needing a `>` or an existing file
+### Changed
+* bumped all dependency versions to latest
+* aliased `cp` to `copy`, `del` to `delete` and `config` to `configure`
+
 ## [0.4.0] - 2020-07-04
 ### Fixed
 [Issue #58](https://github.com/out-of-cheese-error/the-way/issues/58) - changed search highlight and tag colors
@@ -95,6 +114,7 @@ I'll make sure to add changes to it from now, the previous two releases weren't 
 - A first working version of the-way
 - cargo install option
     
+[0.5.0]: https://github.com/out-of-cheese-error/the-way/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/out-of-cheese-error/the-way/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/out-of-cheese-error/the-way/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/out-of-cheese-error/the-way/compare/v0.3.0...v0.3.1